Compare Linelists from NBS and Surveillance Team and Report Differences

Compare Linelists from NBS and Surveillance Team and Report Differences

check_deaths(
  surveillance_file = path_deaths("surv"),
  nbs_file = path_deaths("nbs"),
  directory = "V:/EPI DATA ANALYTICS TEAM/MORTALITY DATA/",
  surveillance_file_id = NULL,
  nbs_file_id = PATIENT_LOCAL_ID,
  save = FALSE
)

Arguments

surveillance_file

The name of the file containing the surveillance team linelist

nbs_file

The name of the file containing the NBS exported linelist

directory

The path to the directory containing both linelists

surveillance_file_id

The NBS ID column in the surveillance file. If NULL, check_deaths() assumes that this is the first column in the file.

nbs_file_id

The NBS ID column in the NBS linelist.

save

Should the results be saved to a file?