Coalesce Report Dates from Archived fst Files

coalesce_report_date() coalesces the output of archive_report_date() into a tibble. Internally, this is a series of join-then-coalesce operations.

coalesce_report_date(
  archive_dir = path_create("V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/",
    "Report Date Linelist/archive/"),
  id = "inv_local_id"
)

Arguments

archive_dir
The location of the archive directory

id
The name of the column to use when joining data; defaults to "inv_local_id"

Value

A tibble containing the id column and a report_date column