Find a Snapshot File and Convert to XLSX Format

convert_snapshot_to_xlsx() finds a snapshot file in directory with date in its name and converts it to an XLSX file of the same name. See convert_to_xlsx for details.

convert_nbs_snapshot(
  date = Sys.Date(),
  directory = paste0("V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/",
    "Sandbox data pull Final/"),
  force = FALSE,
  open = FALSE
)

convert_pcr_snapshot(
  date = Sys.Date(),
  directory = "V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/MSR PCR/",
  force = FALSE,
  open = FALSE
)

convert_snapshot_to_xlsx(
  date = Sys.Date(),
  date_format = "%Y-%m-%d",
  directory,
  force = FALSE,
  open = FALSE
)

Arguments

date
The date included in the name of the file to convert

directory
The directory in which to look for the file to convert

force
Should the file be converted if it will overwrite an existing XLSX file?

open
Should the XLSX file be opened after conversion?

date_format
The format of the date in the file name. The default for the generic function is 'YYYY-MM-DD'; see strptime for details on format specification and the body of convert_pcr_snapshot() for an implementation example (in 'MMDDYYYY' format).

Value

A tidy path to the new XLSX file (invisibly)

Details

The convert_*_snapshot() functions make it easier to convert specific data files.