Cast Lists, Matrices, and More to Source-Specific Data Frames

The as_* family of functions turns existing objects, such as a data frame or matrix, into a tibble of class ael, nbs, or pcr. This is in contrast to the ael(), nbs(), and pcr() functions, which build tibbles from individual columns. Recall that tibble subclasses in coviData have a .date attribute, which assists in handling data made available at different times; you'll need to supply this when using as_ael(), as_nbs(), or as_pcr().

as_ael(
  x,
  ...,
  .date = as.Date(NA),
  .rows = NULL,
  .name_repair = c("check_unique", "unique", "universal", "minimal"),
  rownames = pkgconfig::get_config("tibble::rownames", NULL)
)

as_nbs(
  x,
  ...,
  .date = as.Date(NA),
  .rows = NULL,
  .name_repair = c("check_unique", "unique", "universal", "minimal"),
  rownames = pkgconfig::get_config("tibble::rownames", NULL)
)

as_pcr(
  x,
  ...,
  .date = as.Date(NA),
  .rows = NULL,
  .name_repair = c("check_unique", "unique", "universal", "minimal"),
  rownames = pkgconfig::get_config("tibble::rownames", NULL)
)

Arguments

x
A data frame, list, matrix, or other object that could reasonably be coerced to a tibble.

...
Unused, for extensibility.

.date
A Date indicating when the data was made available. This will usually come from name of the data file; the coviData::load_* family detects this automatically when reading the data. This argument is optional, but omitting it limits the ability of coviData functions to handle data from several reporting dates.

.rows
The number of rows, useful to create a 0-column tibble or just as an additional check.

.name_repair
Treatment of problematic column names:

"minimal": No name repair or checks, beyond basic existence,

"unique": Make sure names are unique and not empty,

"check_unique": (default value), no name repair, but check they are unique,

"universal": Make the names unique and syntactic

a function: apply custom name repair (e.g., .name_repair = make.names for names in the style of base R).

A purrr-style anonymous function, see rlang::as_function()

This argument is passed on as repair to vctrs::vec_as_names(). See there for more details on these terms and the strategies used to enforce them.

rownames
How to treat existing row names of a data frame or matrix:

NULL: remove row names. This is the default.

NA: keep row names.

A string: the name of a new column. Existing rownames are transferred into this column and the row.names attribute is deleted. Read more in rownames.

Details

These functions are thin wrappers around the as_tibble function; see there for details.

See also