Build a Source-Specific Data Frame

ael(), nbs(), and pcr() construct data frames. They are used like tibble(), except an additional .date argument is included.

ael(
  ...,
  .date = NULL,
  .rows = NULL,
  .name_repair = c("check_unique", "unique", "universal", "minimal")
)

nbs(
  ...,
  .date = NULL,
  .rows = NULL,
  .name_repair = c("check_unique", "unique", "universal", "minimal")
)

pcr(
  ...,
  .date = NULL,
  .rows = NULL,
  .name_repair = c("check_unique", "unique", "universal", "minimal")
)

Arguments

...
<dynamic-dots> A set of name-value pairs. These arguments are processed with rlang::quos() and support unquote via !! and unquote-splice via !!!. Use := to create columns that start with a dot.

Arguments are evaluated sequentially. You can refer to previously created elements directly or using the .data pronoun. An existing .data pronoun, provided e.g. inside dplyr::mutate(), is not available.

.date
A Date indicating when the data was made available. This will usually come from name of the data file; the coviData::load_* family detects this automatically when reading the data. This argument is optional, but omitting it limits the ability of coviData functions to handle data from several reporting dates.

.rows
The number of rows, useful to create a 0-column tibble or just as an additional check.

.name_repair
Treatment of problematic column names:

"minimal": No name repair or checks, beyond basic existence,

"unique": Make sure names are unique and not empty,

"check_unique": (default value), no name repair, but check they are unique,

"universal": Make the names unique and syntactic

a function: apply custom name repair (e.g., .name_repair = make.names for names in the style of base R).

A purrr-style anonymous function, see rlang::as_function()

This argument is passed on as repair to vctrs::vec_as_names(). See there for more details on these terms and the strategies used to enforce them.

Value

An object of class ael, nbs, or pcr

Details

These function create data frames of class ael, nbs, or pcr. The data behave identically to a tibble in most functions, but coviData functions uses these classes to treat each data source appropriately. Additionally, the .date argument tells coviData functions when the data was made available, which facilitates automatic handling of data reported at different times.

See also