Save Commonly Used Files in fst Format

Each fst_* function is designed to easily convert a specific file, though they will technically work with any tabular data file.

fst_inv() converts an investigations file

fst_pcr() converts a PCR file

fst_vac() converts a vaccination file

fst_inv(date = NULL, dir = NULL, ext = c("csv", "xlsx"), save_as = NULL, ...)

fst_pcr(date = NULL, save_as = NULL, ...)

fst_vac(path = path_vac(ext = "csv"), save_as = NULL, ...)

Arguments

date
A Date or string in the format "YYYY-MM-DD"

dir
The directory holding the files of interest

ext
The file type to return. Files may be saved as "fst", "csv", or "xlsx".

save_as
Character. The path to save the result fst file. This must have an fst extension; if it does not, the current extension (if any) will be removed and replaced with fst. If left NULL, the path argument is used (with extension replaced by fst).

...
Additional arguments to pass to read_file_delim() or read_file_excel(). The col_types argument is particularly useful to control the type in which the data is saved (default is all character).

path
Character. File path to read from. it will be passed to read_file_excel() if the file extension is "xlsx" or "xls", otherwise it will be passed to read_file_delim()

Value

The dataset being converted, invisibly

Details

For general conversion to fst format, see fst_convert().