Read NBS Files

read_inv() reads investigation data files

read_pct() reads PCR data files

read_inv(
  date = NULL,
  col_select = NULL,
  col_types = vroom::cols(.default = vroom::col_character()),
  ...
)

read_pcr(
  date = NULL,
  col_select = NULL,
  col_types = vroom::cols(.default = vroom::col_character()),
  ...
)

Arguments

date

The download date of the data file to read

col_select

One or more selection expressions, like in dplyr::select(). Use c() or list() to use more than one expression. See ?dplyr::select for details on available selection options.

col_types

One of NULL, a cols() specification, or a string. See vignette("readr") for more details.

If NULL, all column types will be imputed from the first 1000 rows on the input. This is convenient (and fast), but not robust. If the imputation fails, you'll need to increase the guess_max or supply the correct types yourself.

Column specifications created by list() or cols() must contain one column specification for each column. If you only want to read a subset of the columns, use cols_only().

Alternatively, you can use a compact string representation where each character represents one column:

c = character

i = integer

n = number

d = double

l = logical

f = factor

D = date

T = date time

t = time

? = guess

_ or - = skip

By default, reading a file without a column specification will print a message showing what readr guessed they were. To remove this message, set show_col_types = FALSE or set `options(readr.show_col_types = FALSE).

...

Additional arguments to pass to read_file_delim()

Value

A tibble