Read a Tabular Data File

read_file reads delimited- and Excel-type data files. The backend for delimited files is fread; the backend for Excel files is read_excel.

read_file(
  path,
  file_type = c("auto", "delimited", "excel"),
  type_convert = TRUE,
  msg = "Reading file..."
)

Arguments

path
A string indicating the path to the data file; will be expanded if necessary

file_type
A string indicating the type of file to read; only "delimited" and "excel" files are supported. The default is "auto", which determines the file type from the file extension.

type_convert
Should read_file() attempt to guess the data type of the columns after reading?

Note: The default is TRUE for compatibility; want to transition to vroom::vroom() as backend and swap default to FALSE.

msg
A message to be displayed prior to beginning a file read; for use inside other functions

Value

A tibble containing the read data