Replace NBS ID in Surveillance Deaths Linelist with Standardized ID

replace_deaths_id() facilitates comparison of the surveillance deaths linelist with the one exported from NBS by standardizing patient IDs. It saves the result in the file specified in save_as.

replace_deaths_id(
  file = "Working Copy Death  Epi.xlsx",
  directory = "V:/EPI DATA ANALYTICS TEAM/MORTALITY DATA/",
  id = NULL,
  save_as = paste0("sas_data/cleaned_surveillance_copy.xlsx")
)

Arguments

file
The name of the surveillance linelist file

directory
The directory containing the linelist

id
The name of the column specifying the NBS ID in the linelist; if NULL, assumes that this is the first column

save_as
The file path specifying where to save the results

Value

The cleaned data as a tibble (invisibly)