How to Access Sensitive Information Non-Interactively with Renviron

When writing code that might be shared, it's bad practice to hard-code sensitive information in a script (or package, obviously). For data that doesn't need to be shared, but doesn't necessarily need encryption, .Renviron files provide a way to store and retrieve information.

using_Renviron()

Details

A .Renviron file is a special file that you can use to store link[base]{environment variables}, which are just variables that are available using the Sys.getenv function. They're particularly useful for storing information that you don't want to leave your computer, like API keys. You can also store usernames and passwords in this file; it's better than entering them directly, but note that they will not be encrypted. A better solution for usernames, passwords, and other things that should be encrypted is the keyring package. coviData follows this convention; see using_keyring for details on username/ password setup (not yet implemented).

To use the .Renviron file with coviData, follow these steps:

Check your R_USER directory by typing Sys.getenv(R_USER) into the console. We will use the user-level .Renviron file for storage; it lives here.

Go to the R_USER directory in Windows Explorer and create a new text file (right-click -> New -> Text Document). Name this file ".Renviron" and delete the .txt extension.

Open this file using Notepad or another text/code editor

Enter the variables you'd like to store in key=value format. For example, if I wanted to store an API key in the environment variable api_key, I would enter api_key="some_key_to_save" to store it as a string.

Save and exit. Once your restart R, these variables will be available using Sys.env().

(optional) You can hide the .Renviron file in Windows Explorer (right-click -> Properties -> Hidden -> OK) to avoid clutter

coviData currently uses the .Renviron file for REDcap API tokens, but may expand to ESSENCE and NBS tokens as well in the future.