Add Specimen Collection Date to a Report Date Tibble

add_collection_date() reads the specimen collection date column from an NBS data file and adds it to the input data (typically the result of coalesce_report_date()).

add_collection_date(
  .data,
  .from_col = "specimen_coll_dt",
  .id_col = "inv_local_id",
  .from_file = path_create("V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP Data/",
    "Sandbox data pull Final/")
)

Arguments

.data
A tibble returned from coalesce_report_date()

.from_col
A string indicating the variable in .from_file to use as specimen collection date

.id_col
A string indicating the variable in .data and .from_file to use as an ID variable

.from_file
A file path to the NBS data with the desired specimen collection dates; if only a directory is provided, the most recently created NBS file download will be read in